refactor(@leav/ui): make SelectTreeNode autonomous by fetching all data itself
