+++ libs/ui/src/components/SelectTreeNode/_queries/treeDataQuery.graphql
@@ -1,0 +1,8 @@
+query TreeDataQuery($treeId: ID!) {
+    trees(filters: {id: [$treeId]}) {
+        list {
+            id
+            label
+        }
+    }
+}
